+++ core/integration/steps/market/market_config.go
@@ -15,20 +15,22 @@ package market
 import "code.vegaprotocol.io/vega/core/integration/steps/market/defaults"
 
 type Config struct {
-	RiskModels        *riskModels
-	FeesConfig        *feesConfig
-	OracleConfigs     *oracleConfigs
-	PriceMonitoring   *priceMonitoring
-	MarginCalculators *marginCalculators
+	RiskModels          *riskModels
+	FeesConfig          *feesConfig
+	OracleConfigs       *oracleConfigs
+	PriceMonitoring     *priceMonitoring
+	MarginCalculators   *marginCalculators
+	LiquidityMonitoring *liquidityMonitoring
 }
 
 func NewMarketConfig() *Config {
 	unmarshaler := defaults.NewUnmarshaler()
 	return &Config{
-		RiskModels:        newRiskModels(unmarshaler),
-		FeesConfig:        newFeesConfig(unmarshaler),
-		OracleConfigs:     newOracleSpecs(unmarshaler),
-		PriceMonitoring:   newPriceMonitoring(unmarshaler),
-		MarginCalculators: newMarginCalculators(unmarshaler),
+		RiskModels:          newRiskModels(unmarshaler),
+		FeesConfig:          newFeesConfig(unmarshaler),
+		OracleConfigs:       newOracleSpecs(unmarshaler),
+		PriceMonitoring:     newPriceMonitoring(unmarshaler),
+		MarginCalculators:   newMarginCalculators(unmarshaler),
+		LiquidityMonitoring: newLiquidityMonitoring(unmarshaler),
 	}
 }