+++ core/processor/config.go
@@ -13,8 +13,8 @@
 package processor
 
 import (
-	"code.vegaprotocol.io/vega/core/config/encoding"
 	"code.vegaprotocol.io/vega/core/processor/ratelimit"
+	"code.vegaprotocol.io/vega/libs/config/encoding"
 	"code.vegaprotocol.io/vega/logging"
 )