+++ core/validators/topology_eth_key_rotate.go
@@ -198,6 +198,4 @@ func (t *Topology) ethereumKeyRotationBeginBlockLocked(ctx context.Context) {
 	}
 
 	delete(t.pendingPubKeyRotations, t.currentBlockHeight)
-
-	t.tss.changed = true
 }